Create ragnarok_formats crate
This is similar to ragnarok_packets, in that it exposes low level
structs to work with Ragnarok Online content.
