+++ korangar/Cargo.toml
@@ -21,6 +21,7 @@ option-ext = "0.2.0"
 pathfinding = "2.2.2"
 korangar_debug = { workspace = true, optional = true }
 ragnarok_bytes = { workspace = true, features = ["derive", "cgmath"] }
+ragnarok_formats = { workspace = true, features = ["interface"] }
 ragnarok_packets = { workspace = true, features = ["derive", "interface", "packet-to-prototype-element"] }
 rand = "0.8.5"
 random_color = { version = "0.6.1", optional = true }
@@ -32,7 +33,6 @@ serde-xml-rs = "0.6.0"
 tokio = { version = "1.37.0", features = ["full"] }
 vulkano = { git = "https://github.com/vulkano-rs/vulkano.git", rev = "db3df4e55f80c137ea6187250957eb92c2291627" }
 vulkano-shaders = { git = "https://github.com/vulkano-rs/vulkano.git", rev = "db3df4e55f80c137ea6187250957eb92c2291627" }
-vulkano-win = { git = "https://github.com/vulkano-rs/vulkano.git", rev = "db3df4e55f80c137ea6187250957eb92c2291627" }
 walkdir = "2"
 winit = "0.28.7"
 xml-rs = "0.8.0"
