+++ korangar/src/graphics/cameras/player.rs
@@ -1,9 +1,9 @@
 use std::f32::consts::FRAC_PI_2;
 
 use cgmath::{Array, EuclideanSpace, InnerSpace, Matrix4, MetricSpace, Point3, Rad, SquareMatrix, Vector2, Vector3, Vector4};
+use ragnarok_formats::transform::Transform;
 
 use super::{Camera, SmoothedValue};
-use crate::graphics::Transform;
 use crate::interface::layout::{ScreenPosition, ScreenSize};
 
 const ZOOM_SPEED: f32 = 2.0;
