+++ korangar/src/loaders/error.rs
@@ -1,0 +1,10 @@
+use ragnarok_bytes::ConversionError;
+
+use super::FileNotFoundError;
+
+#[derive(Debug)]
+pub enum LoadError {
+    File(FileNotFoundError),
+    Conversion(Box<ConversionError>),
+    UnsupportedFormat(String),
+}
