+++ korangar/src/loaders/mod.rs
@@ -1,8 +1,8 @@
 mod action;
 mod archive;
 pub mod client;
-pub mod color;
 mod effect;
+pub mod error;
 mod font;
 mod gamefile;
 mod map;
@@ -11,18 +11,14 @@ mod script;
 mod server;
 mod sprite;
 mod texture;
-mod version;
 
 pub use self::action::*;
 pub use self::effect::{EffectHolder, EffectLoader, *};
 pub use self::font::{FontLoader, FontSize, Scaling};
 pub use self::gamefile::*;
-#[cfg(feature = "debug")]
-pub use self::map::MapData;
-pub use self::map::{LightSettings, MapLoader, WaterSettings};
+pub use self::map::MapLoader;
 pub use self::model::*;
 pub use self::script::ScriptLoader;
 pub use self::server::{load_client_info, ClientInfo, ServiceId};
 pub use self::sprite::*;
 pub use self::texture::TextureLoader;
-pub use self::version::{InternalVersion, MajorFirst, MinorFirst, Version};
