+++ korangar/src/world/object/mod.rs
@@ -4,6 +4,7 @@ use cgmath::Matrix4;
 use derive_new::new;
 use korangar_interface::elements::PrototypeElement;
 use korangar_interface::windows::PrototypeWindow;
+use ragnarok_formats::transform::Transform;
 use ragnarok_packets::ClientTick;
 
 use crate::graphics::*;
