+++ ragnarok_formats/README.md
@@ -1,0 +1,3 @@
+# Ragnarok Formats
+
+A crate exposing data structures that can be used to parse Ragnarok Online file formats.
