+++ ragnarok_formats/src/color.rs
@@ -1,0 +1,18 @@
+use ragnarok_bytes::ByteConvertable;
+
+#[derive(Clone, Debug, ByteConvertable)]
+#[cfg_attr(feature = "interface", derive(korangar_interface::elements::PrototypeElement))]
+pub struct ColorRGB {
+    pub red: f32,
+    pub green: f32,
+    pub blue: f32,
+}
+
+#[derive(Clone, Debug, ByteConvertable)]
+#[cfg_attr(feature = "interface", derive(korangar_interface::elements::PrototypeElement))]
+pub struct ColorBGRA {
+    pub blue: u8,
+    pub green: u8,
+    pub red: u8,
+    pub alpha: u8,
+}
